Move publishSpotbugsReport to separate file & change spotbugs imports
