+++ Tasks/MavenV3/spotbugsTool/index.ts
@@ -1,0 +1,2 @@
+export * from './spotbugs';
+export * from './publishSpotbugsReport';
