+++ Tasks/MavenV3/spotbugsTool/spotbugs.ts
@@ -1,8 +1,5 @@
 import * as tl from 'azure-pipelines-task-lib/task';
-import { BuildOutput } from 'azure-pipelines-tasks-codeanalysis-common/Common/BuildOutput';
-import { ModuleOutput } from 'azure-pipelines-tasks-codeanalysis-common/Common/ModuleOutput';
-import * as path from 'path';
-import { addPropToJson, copyFile, readXmlFileAsJson, writeJsonAsXmlFile } from './utils';
+import { addPropToJson, copyFile, readXmlFileAsJson, writeJsonAsXmlFile } from '../utils';
 
 /**
  * Gets the plugin nodes and adds it to the original json schema. After that writes the schema to the POM file as XML
@@ -115,36 +112,4 @@ export async function AddSpotbugsPlugin(mavenPOMFile: string): Promise<void> {
         tl.error("Error when updating the POM file: " + err)
         throw err
     }
-}
-
-/**
- * Publishes the spotbugs xml report file to the pipeline artifacts
- * @param mavenPOMFile - Path to the pom.xml configuration file
- * @param buildOutput - Build output from a single or multi module project. Identifies modules based on path conventions.
- */
-export function PublishSpotbugsReport(mavenPOMFile: string, buildOutput: BuildOutput): void {
-    let outputs: ModuleOutput[] = buildOutput.findModuleOutputs();
-    tl.debug(`[CA] ${'Spotbugs'} parser found ${outputs.length} possible modules to upload results from.`);
-
-    const buildRootPath = path.dirname(mavenPOMFile);
-    const reportsPath = path.join(buildRootPath, 'target')
-
-    const reportFile = path.join(reportsPath, 'spotbugsXml.xml')
-
-    const stagingDir: string = path.join(tl.getVariable('build.artifactStagingDirectory'), '.codeAnalysis');
-    const buildNumber: string = tl.getVariable('build.buildNumber');
-
-    const artifactBaseDir: string = path.join(stagingDir, 'CA');
-
-    const destinationDir: string = path.join(artifactBaseDir, outputs[0].moduleName);
-
-    const extension: string = path.extname(reportFile);
-    const reportName: string = path.basename(reportFile, extension);
-
-    const artifactName: string = `${buildNumber}_${reportName}_${'Spotbugs'}${extension}`;
-    copyFile(reportFile, path.join(destinationDir, artifactName));
-
-    tl.command('artifact.upload',
-        { 'artifactname': tl.loc('codeAnalysisArtifactSummaryTitle') },
-        artifactBaseDir);
 }
